Dice: Store u8s instead of size_ts
Don't need 64 bits to store a number from 1-6 :^)

+++ engine/src/Game/Dice.cc
@@ -3,7 +3,7 @@
 #include "Core/Random.h"
 #include "Game/Dice.h"
 
-static auto s_dice_dist = std::uniform_int_distribution<size_t>(k10_DICE_MIN, k10_DICE_MAX);
+static auto s_dice_dist = std::uniform_int_distribution<u8>(k10_DICE_MIN, k10_DICE_MAX);
 
 namespace k10engine::Game {
 
@@ -14,13 +14,13 @@ void Dice::roll()
     m_die_1 = s_dice_dist(rng);
 }
 
-size_t Dice::get_total() const
+u8 Dice::get_total() const
 {
     return m_die_0 + m_die_1;
 }
 
 #ifdef k10_ENABLE_ROLL_DICE_EXACT
-void Dice::set_total(size_t total)
+void Dice::set_total(u8 total)
 {
     assert(1 < total && total < 13);
     if (total > 6) {
