+++ engine/src/Game/Dice.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <cstddef>
 
 namespace k10engine::Game {
 
@@ -15,14 +14,14 @@ class Dice {
 public:
     Dice() {}
     void roll();
-    size_t get_total() const;
+    u8 get_total() const;
 #ifdef k10_ENABLE_ROLL_DICE_EXACT
-    void set_total(size_t);
+    void set_total(u8);
 #endif
 
 private:
-    size_t m_die_0{ 0 };
-    size_t m_die_1{ 0 };
+    u8 m_die_0{ 0 };
+    u8 m_die_1{ 0 };
 };
 
 } // namespace k10engine::Game
